+++ partiql-parser/src/main/kotlin/org/partiql/parser/internal/PartiQLParserDefault.kt
@@ -459,9 +459,9 @@ internal class PartiQLParserDefault : PartiQLParser {
 
             internal val DATE_PATTERN_REGEX = Regex("\\d\\d\\d\\d-\\d\\d-\\d\\d")
 
-            internal val GENERIC_TIME_REGEX = Regex("\\d\\d:\\d\\d:\\d\\d(\\.\\d*)?([+|-]\\d\\d:\\d\\d|Z)?")
+            internal val GENERIC_TIME_REGEX = Regex("\\d\\d:\\d\\d:\\d\\d(\\.\\d*)?([+|-]\\d\\d:\\d\\d)?")
 
-            internal val GENERIC_TIMESTAMP_REGEX = Regex("\\d\\d\\d\\d-\\d\\d-\\d\\d[\\sT]\\d\\d:\\d\\d:\\d\\d(\\.\\d*)?([+|-]\\d\\d:\\d\\d|Z)?")
+            internal val GENERIC_TIMESTAMP_REGEX = Regex("\\d\\d\\d\\d-\\d\\d-\\d\\d\\s\\d\\d:\\d\\d:\\d\\d(\\.\\d*)?([+|-]\\d\\d:\\d\\d)?")
         }
 
         /**
@@ -2073,14 +2073,18 @@ internal class PartiQLParserDefault : PartiQLParser {
             val pattern = ctx.LITERAL_STRING().symbol
             val timeString = ctx.LITERAL_STRING().getStringValue()
             if (GENERIC_TIME_REGEX.matches(timeString).not()) {
-                throw error(pattern, "Expected TIME string to be of the format HH:mm:ss[.SSS][+HH:MM|Z]")
+                throw error(pattern, "Expected TIME string to be of the format HH:mm:ss[.SSS][+HH:MM]")
             }
             val precision = ctx.LITERAL_INTEGER()?.let {
                 val p = it.text.toBigInteger().toInt()
                 if (p < 0 || 9 < p) throw error(it.symbol, "Precision out of bounds [0,9]")
                 p
             }
-            val hasTimezone = timeString.contains(Regex("[+|-]\\d\\d:\\d\\d|Z$")) || ctx.ZONE() != null
+            val hasTimezoneLiteral = timeString.contains(Regex("[+|-]\\d\\d:\\d\\d$"))
+            if (ctx.ZONE() != null && !hasTimezoneLiteral) {
+                throw error(pattern, "TIME WITH TIME ZONE specified but no time offset in literal string")
+            }
+            val hasTimezone = hasTimezoneLiteral || ctx.ZONE() != null
             val type = when (hasTimezone) {
                 true -> {
                     if (precision == null) {
@@ -2104,14 +2108,18 @@ internal class PartiQLParserDefault : PartiQLParser {
             val pattern = ctx.LITERAL_STRING().symbol
             val timestampString = ctx.LITERAL_STRING().getStringValue()
             if (GENERIC_TIMESTAMP_REGEX.matches(timestampString).not()) {
-                throw error(pattern, "Expected TIMESTAMP string to be of the format yyyy-MM-dd HH:mm:ss[.SSS][+HH:MM|Z]")
+                throw error(pattern, "Expected TIMESTAMP string to be of the format yyyy-MM-dd HH:mm:ss[.SSS][+HH:MM]")
             }
             val precision = ctx.LITERAL_INTEGER()?.let {
                 val p = it.text.toBigInteger().toInt()
                 if (p < 0 || 9 < p) throw error(it.symbol, "Precision out of bounds")
                 p
             }
-            val hasTimezone = timestampString.contains(Regex("[+|-]\\d\\d:\\d\\d|Z$")) || ctx.ZONE() != null
+            val hasTimezoneLiteral = timestampString.contains(Regex("[+|-]\\d\\d:\\d\\d$"))
+            if (ctx.ZONE() != null && !hasTimezoneLiteral) {
+                throw error(pattern, "TIMESTAMP WITH TIME ZONE specified but no time offset in literal string")
+            }
+            val hasTimezone = hasTimezoneLiteral || ctx.ZONE() != null
             val type = when (hasTimezone) {
                 true -> {
                     if (precision == null) {